+++ packages/client/src/devices/InputMediaDeviceManager.ts
@@ -8,6 +8,14 @@ import { isReactNative } from '../helpers/platforms';
 export abstract class InputMediaDeviceManager<
   T extends InputMediaDeviceManagerState,
 > {
+  /**
+   * @internal
+   */
+  enablePromise?: Promise<void>;
+  /**
+   * @internal
+   */
+  disablePromise?: Promise<void>;
   constructor(protected readonly call: Call, public readonly state: T) {}
 
   /**
@@ -28,8 +36,14 @@ export abstract class InputMediaDeviceManager<
     if (this.state.status === 'enabled') {
       return;
     }
-    await this.unmuteStream();
-    this.state.setStatus('enabled');
+    this.enablePromise = this.unmuteStream();
+    try {
+      await this.enablePromise;
+      this.state.setStatus('enabled');
+    } catch (error) {
+      this.enablePromise = undefined;
+      throw error;
+    }
   }
 
   /**
@@ -38,12 +52,21 @@ export abstract class InputMediaDeviceManager<
    * @returns
    */
   async disable() {
+    this.state.prevStatus = this.state.status;
     if (this.state.status === 'disabled') {
       return;
     }
-    this.state.prevStatus = this.state.status;
-    await this.muteStream(this.state.disableMode === 'stop-tracks');
-    this.state.setStatus('disabled');
+    this.disablePromise = this.muteStream(
+      this.state.disableMode === 'stop-tracks',
+    );
+    try {
+      await this.disablePromise;
+      this.state.setStatus('disabled');
+      this.disablePromise = undefined;
+    } catch (error) {
+      this.disablePromise = undefined;
+      throw error;
+    }
   }
 
   /**