+++ packages/react-native-sdk/src/components/Call/CallParticipantsList/CallParticipantsList.tsx
@@ -13,7 +13,6 @@ import {
   StreamVideoParticipantPatches,
   VisibilityState,
 } from '@stream-io/video-client';
-import { theme } from '../../../theme';
 import { useDebouncedValue } from '../../../utils/hooks/useDebouncedValue';
 import { useCall } from '@stream-io/video-react-bindings';
 import { ComponentTestIds } from '../../../constants/TestIds';
@@ -239,8 +238,8 @@ const styles = StyleSheet.create({
   },
   participantWrapperHorizontal: {
     // note: if marginHorizontal is changed, be sure to change the width calculation in calculateParticipantViewSize function
-    marginHorizontal: theme.margin.sm,
-    borderRadius: theme.rounded.sm,
+    marginHorizontal: 8,
+    borderRadius: 10,
   },
 });
 
@@ -280,8 +279,8 @@ function calculateParticipantViewSize({
 
   let itemWidth = containerWidth / numberOfColumns;
   if (horizontal) {
-    // in horizontal mode we apply margin to the participant view and that should be subtracted from the width
-    itemWidth = itemWidth - theme.margin.sm * 2;
+    // in horizontal mode we apply margin of 8 to the participant view and that should be subtracted from the width
+    itemWidth = itemWidth - 8 * 2;
   }
 
   return { itemHeight, itemWidth };