+++ packages/react-native-sdk/src/components/Call/index.ts
@@ -3,6 +3,5 @@ export * from './CallTopView';
 export * from './CallLayout';
 export * from './CallControls';
 export * from './CallParticipantsList';
-export * from './IncomingCall';
-export * from './OutgoingCall';
+export * from './RingingCallContent';
 export * from './Lobby';