+++ packages/react-native-sdk/src/theme/icon.ts
@@ -1,9 +1,0 @@
-import { IconScheme } from './types';
-
-export const icon: IconScheme = {
-  xs: { height: 15, width: 15 },
-  sm: { height: 20, width: 20 },
-  md: { height: 25, width: 25 },
-  lg: { height: 30, width: 30 },
-  xl: { height: 35, width: 35 },
-};