+++ packages/react-native-sdk/src/theme/margin.ts
@@ -1,9 +1,0 @@
-import { MarginScheme } from './types';
-
-export const margin: MarginScheme = {
-  xs: 4,
-  sm: 8,
-  md: 16,
-  lg: 24,
-  xl: 32,
-};