+++ sample-apps/react-native/dogfood/src/assets/MicOff.tsx
@@ -1,8 +1,9 @@
 import React from 'react';
 import { Svg, Path } from 'react-native-svg';
+import { ColorValue } from 'react-native';
 
 type Props = {
-  color: string;
+  color: ColorValue;
 };
 
 export const MicOff = ({ color }: Props) => (